+++ crates/nostr-rocksdb/src/lib.rs
@@ -22,7 +22,7 @@ use nostr::nips::nip01::Coordinate;
 use nostr::{Event, EventId, Filter, Timestamp, Url};
 use nostr_database::{
     Backend, DatabaseError, DatabaseIndexes, EventIndexResult, FlatBufferBuilder, FlatBufferDecode,
-    FlatBufferEncode, NostrDatabase, Order, TempEvent,
+    FlatBufferEncode, NostrDatabase, Order,
 };
 use rocksdb::{
     BoundColumnFamily, ColumnFamilyDescriptor, DBCompactionStyle, DBCompressionType, IteratorMode,
@@ -120,7 +120,7 @@ impl RocksDatabase {
             .db
             .full_iterator_cf(&cf, IteratorMode::Start)
             .flatten()
-            .filter_map(|(_, value)| TempEvent::decode(&value).ok())
+            .filter_map(|(_, value)| Event::decode(&value).ok())
             .collect();
 
         // Build indexes